% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{easyRNASeq AnnotParam constructor}
\alias{AnnotParam}
\alias{AnnotParam,GRangesList-method}
\alias{AnnotParam,RangedData-method}
\alias{AnnotParam,character-method}
\alias{AnnotParam,missing-method}
\alias{easyRNASeq AnnotParam constructor}
\title{AnnotParam constructor}
\usage{
\S4method{AnnotParam}{character}(datasource = character(0), type = c("gff3",
  "biomaRt", "gtf", "rda"))
}
\arguments{
\item{datasource}{a character or a \code{\linkS4class{RangedData}} or a \code{\linkS4class{GRangesList}} object. See details.}

\item{type}{one of NULL, biomaRt, gff3, gtf or rda. Default to NULL. See details.}
}
\description{
This constructs a \code{\linkS4class{AnnotParam}} object.
The datasource parameter (see details) is mandatory, however
other parameters, \emph{i.e.} when the datasource is not a
\code{\linkS4class{GRangesList}} or \code{\linkS4class{RangedData}}
default to "genes" and gff3", indicating that the datasource is in
the gff3 format and that the contained information needs to be grouped
by "genes". This representing the most common use case. Hence, it is
left to the user to refine the parameters accordingly to the annotation
he is providing or whishes to retrieve.
}
\details{
Note that calling the constructor without argument fails, as the datasource
is a mandatory parameter. Calling the constructor with additional (not all)
parameters will affect the value of the selected parameters, leaving the other
parameters unaffected.
There are three parameters for an \code{\linkS4class{AnnotParam}}
object:
\itemize{
\item{datasource}{If no type is provided, the datasource should
be either a \code{\linkS4class{GRangesList}}(prefered) or a
\code{\linkS4class{RangedData}} (subject to future deprecation)
object  containing the genic information. These can be obtained
using the \code{\link[easyRNASeq:easyRNASeq-annotation-methods]{getAnnotation}} function.}
\item{type}{One of biomaRt, gff3, gtf or rda. The default is "gff3".
In all cases, the datasource is a \code{character} describing:
\itemize{
\item For biomaRt, the name of the organism as known by the ensembl Mart,
\emph{e.g.} dmelanogaster or hsapiens.
\item For gff3, gtf or rda, the filename (including the full or relative path).
}
}
}
}
\examples{
## create an object to retrieve annotation from biomaRt
annotParam <- AnnotParam(datasource="Hsapiens",type="biomaRt")

## get the datasource and type
datasource(annotParam)
type(annotParam)

## create an object to retrieve annotation from an rda object
\dontrun{
library(RnaSeqTutorial)
annotParam <- AnnotParam(datasource=system.file(
  package="RnaSeqTutorial",
  "data","gAnnot.rda"
),type="rda")
}
}
\seealso{
\itemize{
\item \code{\linkS4class{GRangesList}}
\item \code{\linkS4class{RangedData}}
\item \code{\link[easyRNASeq:easyRNASeq-annotation-methods]{getAnnotation}}
}
}

