\docType{class}
\name{BamParam class}
\alias{BamParam class}
\alias{BamParam-class}
\title{Class "BamParam"}
\description{
A class describing the parameters of a bam file issued from
an RNA-Seq experiment.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{new("BamParam", ...)} or using the BamParam
  constructor.
}
\examples{
showClass("BamParam")
}
\author{
Nicolas Delhomme
}
\seealso{
\itemize{ \item \code{\linkS4class{RnaSeqParam}} \item
\code{\link[easyRNASeq:easyRNASeq-RnaSeqParam]{RnaSeqParam
constructor}} ##\item
\code{\link[easyRNASeq:easyRNASeq-RnaSeqParam-accessors]{RnaSeqParam
accessors}} \item
\code{\link[easyRNASeq:easyRNASeq-simpleRNASeq]{simpleRNASeq
function}} \item \code{\linkS4class{AnnotParam}} \item
\code{\link[easyRNASeq:easyRNASeq-AnnotParam]{AnnotParam
constructor}} }
}
\keyword{classes}

