% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RNAseq class}
\alias{RNAseq class}
\alias{RNAseq-class}
\alias{RNAseq}
\title{Class "RNAseq"}
\description{
A class holding all the necessary information and annotation to summarize
couts (number of reads) per features (i.e. exons or transcripts or genes)
for RNA-Seq experiments.
}
\section{Objects from the Class}{
 Objects can be created by calls of the
form \code{new("RNAseq", ...)}.
}

\examples{

showClass("RNAseq")

}
\seealso{
\itemize{ \item \code{\linkS4class{GRangesList}}
\item \code{\linkS4class{RleList}}
\item \code{\link[easyRNASeq:easyRNASeq-easyRNASeq]{easyRNASeq}}
\item \code{\link[easyRNASeq:easyRNASeq-accessors]{easyRNASeq accessors}}
\item \code{\link[easyRNASeq:easyRNASeq-annotation-methods]{easyRNASeq annotation}}
\item \code{\link[easyRNASeq:easyRNASeq-correction-methods]{easyRNASeq correction (FPKM)}}
\item \code{\link[easyRNASeq:easyRNASeq-coverage-methods]{easyRNASeq coverage}}
\item \code{\link[easyRNASeq:easyRNASeq-summarization-methods]{easyRNASeq summarization}}
\item \code{\link[easyRNASeq:print-methods]{easyRNASeq print methods}}
}
}
\author{
Nicolas Delhomme
}
\keyword{classes}
