% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRNASeq-package.R
\name{easyRNASeq-datasets}
\alias{easyRNASeq-datasets}
\alias{RobinsonDelhomme2014}
\title{Dataset included in the package}
\description{
The package contains a dataset from the \emph{Robinson, Delhomme et al., 2014}
publication.
\itemize{
\item{RobinsonDelhomme2014}{a normalised expression count table. This dataset was
generated from 17 \emph{Populus tremula} - Eurasian
aspen - trees used to assess the sexual dimorphism of this dioecious species. This
count matrix has been generating following published pre-processing
guidelines - see \url{http://www.epigenesys.eu/en/protocols/bio-informatics/1283-guidelines-for-rna-seq-data-analysis} -
and the resulting HTSeq files have been collated and the obtained raw count
matrix submitted to a variance stabilising transformation. Subsequently, the
values have been transformed so that the minimal vst values - that
corresponds to an absence of expression - is 0. Hence the counts in the
matrix are library-size normalized, variance stabilised expression values,
with a minimal value of 0.}
}
}
\keyword{data}
