% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRNASeq-annotation.R
\name{easyRNASeq defunct annotation methods}
\alias{easyRNASeq defunct annotation methods}
\alias{fetchAnnotation-defunct}
\alias{knownOrganisms-defunct}
\title{Defunct annotation function}
\description{
The \code{fetchAnnotation} and \code{knownOrganisms} function are now
defunct. The \code{fetchAnnotation} function has been replaced by the
\code{\link[easyRNASeq:easyRNASeq-annotation-methods]{getAnnotation}} method.
}
\author{
Nicolas Delhomme
}

