% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRNASeq-package.R
\name{easyRNASeq-global-variables}
\alias{easyRNASeq-global-variables}
\alias{.onAttach}
\alias{ANNOTATION.TYPE}
\alias{GTF.FIELDS}
\alias{TUTORIAL.DATA}
\title{Objects created when the package is attached.}
\arguments{
\item{libname}{a character string giving the library directory where the package defining the namespace was found.}

\item{pkgname}{a character string giving the name of the package.}
}
\description{
The package creates the following objects when attached
\itemize{
\item GTF.FIELDS
\item ANNOTATION.TYPE
\item TUTORIAL.DATA
}
}
\details{
These objects hold the following information
\itemize{
\item GTF.FIELDS \code{c("gene_id","transcript_id","exon_id","gene_name")}
\item ANNOTATION.TYPE \code{c(mRNA="mRNA",exon="exon")}
\item TUTORIAL.DATA: The list of files needed for the vignette, help and test
pages
}
and are designed as global variables to expose the
fact that they are hardcoded. These exist as
placeholder in case a user would require different
values for these.
}
\seealso{
\code{\link[base:ns-hooks]{.onAttach}} in the \code{base} package.
}
\keyword{internal}
