% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{easyRNASeq RnaSeqParam constructor}
\alias{RnaSeqParam}
\alias{RnaSeqParam,ANY-method}
\alias{easyRNASeq RnaSeqParam constructor}
\title{RnaSeqParam constructor}
\usage{
\S4method{RnaSeqParam}{ANY}(annotParam = AnnotParam(),
  bamParam = BamParam(), countBy = c("exons", "features", "genes",
  "transcripts"), precision = c("read", "bp"))
}
\arguments{
  \item{annotParam}{An object derived from class
  \code{AnnotParam}.}

  \item{bamParam}{An object derived from class
  \code{BamParam}.}

  \item{countBy}{TODO}

  \item{precision}{A character value, either 'read' or 'bp'
  that defines the precision at which counting is done,
  either per read or per covered bp. 'read' is the
  default.}
}
\description{
TODO ALL WRONG!!! This constructs a
\code{\linkS4class{RnaSeqParam}} object. The default
parameters are derived from the currently most common
RNA-Seq experimental use-case and are detailed below:
\itemize{ \item paired is TRUE, \emph{i.e.} paired-end
sequencing is expected. \item stranded is FALSE \emph{i.e.}
stranded sequencing is not expected. \item yieldSize is set
to 100,000. This is the amount of reads iteratively
processed from the bam file stream. It is a compromise
between speed, process-parallelization and memory usage. }
}
\details{
Calling the constructor without argument result in the
default parameter described above to be returned. Calling
the constructor with any parameter will affect the value of
the selected parameters, leaving the other parameters
unaffected.
}
\examples{
annotParam <- AnnotParam(
                datasource=system.file(
                                "extdata",
                                "Dmel-mRNA-exon-r5.52.gff3",
                                package="RnaSeqTutorial"))

## create the RnaSeqParam
rsp <- RnaSeqParam(annotParam=annotParam)

## change some defaults
RnaSeqParam(countBy="features",annotParam=annotParam)
RnaSeqParam(bamParam=BamParam(stranded=TRUE,yieldSize=1L),annotParam=annotParam)
}

