% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{IRanges additional methods}
\alias{IRanges additional methods}
\alias{ranges}
\alias{ranges,RNAseq-method}
\title{Extension of the IRanges package}
\usage{
\S4method{ranges}{RNAseq}(x)
}
\arguments{
  \item{x}{An object of the \code{\linkS4class{RNAseq}}
  class}
}
\value{
An \code{\linkS4class{IRangesList}} object, where the split
is performed by seqnames (\emph{e.g.} chromosomes).
}
\description{
Return the ranges of the genomic annotation.
}
\details{
It retrieves the object stored in the genomicAnnotation
slot of the RNAseq object and apply the \code{ranges}
function on it. The object retrieved can be of the
\code{\linkS4class{RangedData}} or
\code{\linkS4class{GRangesList}} class.
}
\examples{
\dontrun{
	library("RnaSeqTutorial")
	obj <- new('RNAseq',
		organismName="Dmelanogaster",
		readLength=36L,
		chrSize=as.list(seqlengths(Dmelanogaster))
		)

	obj <- fetchAnnotation(obj,
				method="gff",
                                filename=system.file(
						"extdata",
						"annot.gff",
						package="RnaSeqTutorial"))
	ranges(obj)
	}
}
\author{
Nicolas Delhomme
}
\keyword{methods}

