% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parallel-methods.R
\docType{methods}
\name{parallel additional methods}
\alias{parallel additional methods}
\alias{parallelize}
\alias{parallelize,BamFileList,function-method}
\alias{parallelize,GRangesList,function-method}
\alias{parallelize,list,function-method}
\alias{parallelize,vector,function-method}
\title{parallel additional methods}
\usage{
\S4method{parallelize}{list,`function`}(obj = list(), fun = NULL,
  nnodes = 1, ...)
}
\arguments{
  \item{obj}{the object which processing has to be
  parallelizes}

  \item{fun}{the function to be applied in parallel}

  \item{nnodes}{the number of nodes to use}

  \item{...}{additional arguments passed to the function
  \code{fun}}
}
\value{
the result of the
\code{\link[parallel:clusterApply]{clusterApply}} function.
}
\description{
Functions defined in the easyRNASeq package that enhance
the parallel package.
}
\details{
The parallelize function ease the use of the parallel
package. If the number of nodes provided by the user is 1,
then a simple 'lapply' is used, otherwise a cluster object
is created and the object dispatched for parallelization.
}
\examples{
parallelize(list(a<-c(1,2),b<-c(2,1)),sum,nnodes=1)
}
\author{
Nicolas Delhomme
}
\seealso{
\code{\link[parallel:clusterApply]{clusterApply}}
\code{\link[parallel:makePSOCKcluster]{makePSOCKcluster}}
\code{\link[parallel:stopCluster]{stopCluster}}
}
\keyword{methods}

