\name{print-methods}
\docType{methods}
\alias{print}

\title{ Method to print a RNAseq object}

\description{
  Print information about a \code{\linkS4class{RNAseq}} object.
}

\arguments{
  \item{rnaSeq}{An object derived from class \code{\linkS4class{RNAseq}}}
  \item{verbose}{A logical to have a verbose or not output. Default to FALSE}
  \item{\dots}{Additional arguments, currently unused.}
}

\value{
  Print information about a \code{\linkS4class{RNAseq}} object.
}

\author{Nicolas Delhomme}
\keyword{methods}

