\name{show-methods}
\docType{methods}
\alias{show-methods}

\title{ Display the content of a RNAseq object}
\description{
	Display the content of a \code{\linkS4class{RNAseq}} object.
}
\section{Methods}{
\describe{
	\item{\code{signature(object = "RNAseq")}}{
		Display the values of the different slots of the \code{\linkS4class{RNAseq}} object.
	}
}}
\keyword{methods}

