% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/easyRNASeq-BamParam-accessors.R
\docType{methods}
\name{easyRNASeq BamParam accessors}
\alias{BamParam-accessors}
\alias{easyRNASeq BamParam accessors}
\alias{paired}
\alias{paired,BamParam-method}
\alias{stranded}
\alias{stranded,BamParam-method}
\alias{yieldSize}
\alias{yieldSize,BamParam-method}
\title{Accessors for BamParam class}
\usage{
yieldSize(object,...)
paired(object)
stranded(object)
}
\arguments{
  \item{object}{An object derived from class
  \code{BamParam}.}

  \item{...}{Additional parameter inherited from the
  \code{\link[Rsamtools:RsamtoolsFile-class]{Rsamtools
  package yieldSize function}}. Ignored here.}
}
\value{
The value of the corresponding slot.
}
\description{
These functions and generics define `accessors` (to get and
set values) for \code{\linkS4class{BamParam}} objects
within the \pkg{easyRNASeq} package.
}
\examples{
bp <- BamParam()
## get the yieldSize Parameter
ysize <-yieldSize(bp)
}
\author{
Nicolas Delhomme
}
\seealso{
The
\code{\link[easyRNASeq:easyRNASeq-BamParam-class]{BamParam}}
class The
\code{\link[easyRNASeq:easyRNASeq-RnaSeqParam-accessors]{RnaSeqParam
yieldSize}} accessor
}
\keyword{manip}

