% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RNAseq class}
\alias{RNAseq class}
\alias{RNAseq-class}
\alias{RNAseq}
\title{Class "RNAseq"}
\description{
A class holding all the necessary information and annotation to summarize
couts (number of reads) per features (i.e. exons or transcripts or genes)
for RNA-Seq experiments.
}
\section{Objects from the Class}{
 Objects can be created by calls of the
form \code{new("RNAseq", ...)}.
}

\examples{

showClass("RNAseq")

}
\seealso{
\itemize{ \item\code{\linkS4class{GRangesList}}
\item\code{\linkS4class{RleList}}
\item\code{\link{easyRNASeq-easyRNASeq}}
\item\code{\link{easyRNASeq-accessors}}
\item\code{\link{easyRNASeq-annotation-methods}}
\item\code{\link{easyRNASeq-correction-methods}}
\item\code{\link{easyRNASeq-coverage-methods}}
\item\code{\link{easyRNASeq-summarization-methods}}
\item\code{\link{print-methods}}
}
}
\author{
Nicolas Delhomme
}
\keyword{classes}
