% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RNAseq class}
\alias{RNAseq}
\alias{RNAseq class}
\alias{RNAseq-class}
\title{Class "RNAseq"}
\description{
A class holding all the necessary information and annotation to summarize
couts (number of reads) per features (i.e. exons or transcripts or genes)
for RNA-Seq experiments.
}
\section{Objects from the Class}{
 Objects can be created by calls of the
form \code{new("RNAseq", ...)}.
}
\examples{
showClass("RNAseq")
}
\author{
Nicolas Delhomme
}
\seealso{
\itemize{ \item\code{\linkS4class{RangedData}}
\item\code{\linkS4class{RleList}}
\item\code{\link[easyRNASeq:easyRNASeq-easyRNASeq]{easyRNASeq function}}
\item\code{\link[easyRNASeq:easyRNASeq-accessors]{RNAseq accessors}}
\item\code{\link[easyRNASeq:easyRNASeq-annotation-methods]{easyRNASeq
annotation methods}}
\item\code{\link[easyRNASeq:easyRNASeq-correction-methods]{easyRNASeq
correction methods}}
\item\code{\link[easyRNASeq:easyRNASeq-coverage-methods]{easyRNASeq
coverage methods}}
\item\code{\link[easyRNASeq:easyRNASeq-summarization-methods]{easyRNASeq
summarization methods}} \item\code{\link[easyRNASeq:print-methods]{print}}
}
}
\keyword{classes}

