% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/easyRNASeq-methods.R
\docType{methods}
\name{count,character-method}
\alias{count}
\alias{count,character-method}
\title{count method}
\usage{
\S4method{count}{character}(filesDirectory = getwd(),
  outputFormat = "SummarizedExperiment", ...)
}
\arguments{
  \item{filesDirectory}{The directory where the files to be
  used are located.}

  \item{outputFormat}{By default, easyRNASeq returns a
  \code{\linkS4class{RangedSummarizedExperiment}} object. If one of
  \code{DESeq},\code{edgeR},\code{RNAseq}, \code{matrix} is
  provided then the respective object is returned. Ideally,
  this option should get deprecated and only a
  RangedSummarizedExperiment object returned.}

  \item{...}{currently additional arguments to the
  easyRNASeq function.}
}
\value{
Returns a \code{\linkS4class{RangedSummarizedExperiment}} object. If the
\code{outputFormat} option has been set, a corresponding
object is returned: a count table (a matrix of m features x
n samples), a
\code{\link[DESeq:newCountDataSet]{DESeq:newCountDataset}},
a \code{\link[edgeR:DGEList]{edgeR:DGEList}} or
\code{\linkS4class{RNAseq}}.
}
\description{
This function is to supersed the easyRNASeq function in
order to consolidate the option parameters as well as the
option output. Ideally, the only output would be a
RangedSummarizedExperiment object.
}
\examples{
\dontrun{
	library("RnaSeqTutorial")
	library(BSgenome.Dmelanogaster.UCSC.dm3)

	## creating a count table from 4 bam files
	sumExp <- count(filesDirectory=system.file(
                           "extdata",
                           package="RnaSeqTutorial"),
                         pattern="[A,C,T,G]{6}\\\\.bam$",
                         readLength=30L,
                         organism="Dmelanogaster",
                         chr.sizes=seqlengths(Dmelanogaster),
                         annotationMethod="rda",
                         annotationFile=system.file(
                           "data",
                           "gAnnot.rda",
                           package="RnaSeqTutorial"),
                         count="exons"
                         )
    ## the counts
    assays(sumExp)
    ## the sample info
    colData(sumExp)
    ## the 'features' info
    rowRanges(sumExp)
}
}
\author{
Nicolas Delhomme
}
\seealso{
\code{\linkS4class{RNAseq}}
\code{\linkS4class{RangedSummarizedExperiment}}
\code{\link[edgeR:DGEList]{edgeR:DGEList}}
\code{\link[DESeq:newCountDataSet]{DESeq:newCountDataset}}
\code{\link[easyRNASeq:easyRNASeq-annotation-methods]{easyRNASeq:knownOrganisms}}
\code{\link[ShortRead:readAligned]{ShortRead:readAligned}}
}
\keyword{methods}

