% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Defunct functions}
\alias{Defunct functions}
\alias{plotDispersionEstimates,DGEList-method}
\title{The following function are defunct:
\itemize{
\item \code{lotDispersionEstimates,DGEList-method}
}}
\description{
\itemize{
\item The \code{plotDispersionEstimates,DGEList-method}
function is superseded by the \code{\link[edgeR:plotBCV]{plotBCV}} function
as the \pkg{edgeR} DGEList object structure changed
}
}

