% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/easyRNASeq-package.R
\docType{package}
\name{easyRNASeq package}
\alias{BamFileList}
\alias{BamFileList-class}
\alias{IRanges}
\alias{RangedData}
\alias{RangedData-class}
\alias{SRFilterResult}
\alias{RangedSummarizedExperiment-class}
\alias{alignData}
\alias{assay}
\alias{chromosomeFilter}
\alias{compose}
\alias{easyRNASeq package}
\alias{easyRNASeq package-package}
\alias{easyRNASeq-package}
\alias{nFilter}
\alias{type}
\title{Count summarization and normalization pipeline for Next Generation Sequencing data.}
\description{
Offers functionalities to summarize read counts per feature
of interest, e.g. exons, transcripts, genes, etc. Offers
functionalities to normalize the summarized counts using
3rd party packages like
\code{\link[DESeq:newCountDataSet]{DESeq}} or
\code{\link[edgeR:DGEList]{edgeR}}.
}
\details{
\tabular{ll}{ Package: \tab easyRNASeq\cr Type: \tab
Package\cr Version: \tab 2.3.4\cr Date: \tab 2015-03-16\cr
License: \tab Artistic-2.0\cr LazyLoad: \tab yes\cr
Depends: \tab methods, parallel, Biobase, BiocGenerics,
biomaRt, Biostrings, edgeR, DESeq, genomeIntervals,
GenomeInfoDb, GenomicAlignments, GenomicRanges, graphics,
IRanges, LSD, Rsamtools, S4Vectors, ShortRead, utils\cr
Suggests: \tab BSgenome.Dmelanogaster.UCSC.dm3 }
}
\section{Methods}{
  The main function
  \code{\link[easyRNASeq:easyRNASeq-easyRNASeq]{easyRNASeq}}
  will summarize the counts per feature of interest, for as
  many samples as provided and will return a count matrix
  (N*M) where N are the features and M the samples. This
  data can be corrected to \pkg{RPKM} in which case a
  matrix of corrected value is returned instead, with the
  same dimensions. Alternatively a
  \code{\linkS4class{RangedSummarizedExperiment}} can be returned
  and this is expected to be the default in the upcoming
  version of easyRNASeq (as of 1.5.x). If the necessary
  sample information are provided, the data can be
  normalized using either
  \code{\link[DESeq:newCountDataSet]{DESeq}} or
  \code{\link[edgeR:DGEList]{edgeR}} and the corresponding
  package object returned. For more insider details, and
  step by step functions, see: \tabular{ll}{
  \code{\link[easyRNASeq:ShortRead-methods]{ShortRead
  methods}} for pre-processing the data.
  \code{\link[easyRNASeq:easyRNASeq-annotation-methods]{easyRNASeq
  annotation methods}} for getting the annotation.
  \code{\link[easyRNASeq:easyRNASeq-coverage-methods]{easyRNASeq
  coverage methods}} for computing the coverage from a
  Short Read Alignment file.
  \code{\link[easyRNASeq:easyRNASeq-summarization-methods]{easyRNASeq
  summarization methods}} for summarizing the data.
  \code{\link[easyRNASeq:easyRNASeq-correction-methods]{easyRNASeq
  correction methods}} for correcting the data (i.e.
  generating RPKM).
  \code{\link[easyRNASeq:edgeR-methods]{edgeR methods}} for
  post-processing the data.
  \code{\link[easyRNASeq:DESeq-methods]{DESeq methods}} for
  post-processing the data.  }
}
\examples{
\dontrun{
	library("RnaSeqTutorial")
	library(BSgenome.Dmelanogaster.UCSC.dm3)

	## creating a count table from 4 bam files
	count.table <- easyRNASeq(filesDirectory=
		    			system.file(
					"extdata",
					package="RnaSeqTutorial"),
					pattern="[A,C,T,G]{6}\\\\.bam$",
				format="bam",
				readLength=36L,
				organism="Dmelanogaster",
				chr.sizes=as.list(seqlengths(Dmelanogaster)),
				annotationMethod="rda",
				annotationFile=system.file(
				                            "data",
							    "gAnnot.rda",
							    package="RnaSeqTutorial"),
				count="exons")

	}
}
\author{
Nicolas Delhomme, Bastian Schiffthaler, Ismael Padioleau
}
\seealso{
The class RNAseq specification: \code{\linkS4class{RNAseq}}

The default output class specification:
\code{\linkS4class{RangedSummarizedExperiment}}

The imported packages:
\code{\link[biomaRt:useMart]{biomaRt}}
\code{\link[edgeR:DGEList]{edgeR}}
\code{\link[genomeIntervals:Genome_intervals_stranded-class]{genomeIntervals}}
\code{\link[Biostrings:XString-class]{Biostrings}}
\code{\link[BSgenome:BSgenome-class]{BSgenome}}
\code{\link[DESeq:newCountDataSet]{DESeq}}
\code{\link[GenomicRanges:GRanges-class]{GenomicRanges}}
\code{\link[IRanges:IRanges-constructor]{IRanges}}
\code{\link[Rsamtools:scanBam]{Rsamtools}}
\code{\link[ShortRead:readAligned]{ShortRead}}

The suggested packages:
\code{\link[parallel:makeCluster]{parallel}}
\code{\link[GenomicFeatures:TxDb-class]{GenomicFeatures}}

The following classes and functions that are made available
from other packages: \itemize{ \item{Classes}{
\code{\linkS4class{BamFileList}}
\code{\linkS4class{RangedData}}
\code{\linkS4class{RangedSummarizedExperiment}} }
\item{Functions/Methods}{
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{ The
RangedSummarizedExperiment assay accessor} }
\code{\link[Rsamtools:BamFileList]{The BamFileList
constructor}} \code{\link[IRanges:IRanges-constructor]{The
IRanges constructor}}
\code{\link[IRanges:RangedData-class]{The RangedData
constructor}} \code{\link[ShortRead:srFilter]{For the
SRFilterResult, chromosomeFilter, compose and nFilter
methods}} } }
}
\keyword{package}

