% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genomeIntervals-methods.R
\docType{methods}
\name{genomeIntervals additional methods}
\alias{genomeIntervals additional methods}
\alias{type,Genome_intervals-method}
\title{Extension for the genomeIntervals package}
\usage{
\S4method{type}{Genome_intervals}(x)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{Genome_intervals}}}
}
\value{
\describe{
\item{type}{ The content of the
type column, usually a factor or a character vector } }
}
\description{
\describe{
\item{type}{ Another way to access the content of the gff type
column.  }}
}
\examples{
\dontrun{
annot<-readGff3(system.file("extdata",
                            "Dmel-mRNA-exon-r5.52.gff3",
                            package="RnaSeqTutorial")
type(annot)
}
}
\author{
Nicolas Delhomme
}
\seealso{
\itemize{
\item \code{\link[genomeIntervals:Genome_intervals_stranded-class]{genomeIntervals
object}}
\item \code{\link[genomeIntervals:readGff3]{readGff3 function}}
}
}

