% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeIntervals-methods.R
\name{genomeIntervals additional methods}
\alias{genomeIntervals additional methods}
\alias{type,Genome_intervals-method}
\title{Extension for the genomeIntervals package}
\usage{
\S4method{type}{Genome_intervals}(x)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{Genome_intervals}}}
}
\value{
\describe{
\item{type}{ The content of the
type column, usually a factor or a character vector } }
}
\description{
\describe{
\item{type}{ Another way to access the content of the gff type
column.  }}
}
\examples{
# library
library(genomeIntervals)

# fetch the example data
gffFilePath <- fetchData("Dmel-mRNA-exon-r5.52.gff3.gz")
annot<-readGff3(gffFilePath,quiet=TRUE)
type(annot)

}
\seealso{
\itemize{
\item \code{\link[genomeIntervals:Genome_intervals_stranded-class]{genomeIntervals
object}}
\item \code{\link[genomeIntervals]{genomeIntervals-readGff3}}
}
}
\author{
Nicolas Delhomme
}
