% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeIntervals-methods.R
\docType{methods}
\name{genomeIntervals additional methods}
\alias{genomeIntervals additional methods}
\alias{type,Genome_intervals-method}
\title{Extension for the genomeIntervals package}
\usage{
\S4method{type}{Genome_intervals}(x)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{Genome_intervals}}}
}
\value{
\describe{
\item{type}{ The content of the
type column, usually a factor or a character vector } }
}
\description{
\describe{
\item{type}{ Another way to access the content of the gff type
column.  }}
}
\examples{

library(curl)
library(genomeIntervals)
invisible(curl_download(paste0("https://github.com/UPSCb/UPSCb/raw/",
                               "master/tutorial/easyRNASeq/Dmel-mRNA-exon-r5.52.gff3.gz"),
                        "Dmel-mRNA-exon-r5.52.gff3.gz"))
annot<-readGff3("Dmel-mRNA-exon-r5.52.gff3.gz",quiet=TRUE)
type(annot)

}
\seealso{
\itemize{
\item \code{\link[genomeIntervals:Genome_intervals_stranded-class]{genomeIntervals
object}}
\item \code{\link[genomeIntervals:readGff3]{readGff3 function}}
}
}
\author{
Nicolas Delhomme
}
