% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\docType{methods}
\name{print methods}
\alias{print methods}
\alias{print}
\alias{print,RNAseq-method}
\alias{print,AnnotParam-method}
\alias{print,BamParam-method}
\alias{print,RnaSeqParam-method}
\title{Pretty print the content of classes from the easyRNASeq package.}
\usage{
\S4method{print}{RNAseq}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{An object from class \code{\linkS4class{RNAseq}},
\code{\linkS4class{AnnotParam}},
\code{\linkS4class{BamParam}} or \code{\linkS4class{RnaSeqParam}}}

\item{verbose}{A logical to have a verbose or not output. Default to FALSE
For object of the \code{\linkS4class{RNAseq}} class only.}

\item{...}{Additional arguments, currently unused.}
}
\value{
Print information about the provided object.
}
\description{
Print information about a \code{\linkS4class{RNAseq}},
\code{\linkS4class{AnnotParam}}, \code{\linkS4class{BamParam}} or
\code{\linkS4class{RnaSeqParam}} object.
}
\author{
Nicolas Delhomme
}
\keyword{methods}
