\docType{methods}
\name{print methods}
\alias{print}
\alias{print methods}
\alias{print,RNAseq-method}
\title{Method to print a RNAseq object}
\usage{
\S4method{print}{RNAseq}(x, verbose = FALSE, ...)
}
\arguments{
  \item{x}{An object derived from class
  \code{\linkS4class{RNAseq}}}

  \item{verbose}{A logical to have a verbose or not output.
  Default to FALSE}

  \item{...}{Additional arguments, currently unused.}
}
\value{
Print information about a \code{\linkS4class{RNAseq}}
object.
}
\description{
Print information about a \code{\linkS4class{RNAseq}}
object.
}
\author{
Nicolas Delhomme
}
\keyword{methods}

