\docType{methods}
\name{show methods}
\alias{show methods}
\alias{show,RNAseq-method}
\title{Display the content of a RNAseq object}
\usage{
\S4method{show}{RNAseq}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
Display the content of a \code{\linkS4class{RNAseq}}
object.
}
\section{Methods}{
  \describe{ \item{list("signature(object = \"RNAseq\")")}{
  Display the values of the different slots of the
  \code{\linkS4class{RNAseq}} object.  } }
}
\keyword{methods}

